# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Journalists Service",
    "description": "Manages journalist profiles, outlet affiliations, email discovery pipeline, and engagement tracking.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3020"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "Journalist": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "entityType": {
            "type": "string",
            "enum": [
              "individual",
              "organization"
            ]
          },
          "journalistName": {
            "type": "string"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "entityType",
          "journalistName",
          "firstName",
          "lastName",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ValidJournalistEmail": {
        "type": "object",
        "properties": {
          "journalistId": {
            "type": "string",
            "format": "uuid"
          },
          "outletId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "isValid": {
            "type": "boolean"
          },
          "type": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "confidence": {
            "type": "number"
          }
        },
        "required": [
          "journalistId",
          "outletId",
          "email",
          "isValid",
          "type",
          "source",
          "confidence"
        ]
      },
      "HuntedEmailEvent": {
        "type": "object",
        "properties": {
          "outletId": {
            "type": "string",
            "format": "uuid"
          },
          "journalistId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "huntedAt": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "nullable": true
          },
          "acceptAll": {
            "type": "boolean",
            "nullable": true
          }
        },
        "required": [
          "outletId",
          "journalistId",
          "email",
          "huntedAt",
          "status",
          "score",
          "acceptAll"
        ]
      },
      "HuntedIndividualEvent": {
        "type": "object",
        "properties": {
          "outletId": {
            "type": "string",
            "format": "uuid"
          },
          "journalistId": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "huntedAt": {
            "type": "string"
          },
          "position": {
            "type": "string",
            "nullable": true
          },
          "verificationStatus": {
            "type": "string",
            "nullable": true
          },
          "score": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "outletId",
          "journalistId",
          "firstName",
          "lastName",
          "domain",
          "huntedAt",
          "position",
          "verificationStatus",
          "score"
        ]
      },
      "SearchedEmailEvent": {
        "type": "object",
        "properties": {
          "outletId": {
            "type": "string",
            "format": "uuid"
          },
          "journalistId": {
            "type": "string",
            "format": "uuid"
          },
          "journalistEmail": {
            "type": "string"
          },
          "searchedAt": {
            "type": "string"
          },
          "sourceStatus": {
            "type": "string",
            "nullable": true
          },
          "sourceQuote": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "outletId",
          "journalistId",
          "journalistEmail",
          "searchedAt",
          "sourceStatus",
          "sourceQuote"
        ]
      },
      "NeedEmailUpdate": {
        "type": "object",
        "properties": {
          "outletId": {
            "type": "string",
            "format": "uuid"
          },
          "journalistId": {
            "type": "string",
            "format": "uuid"
          },
          "journalistName": {
            "type": "string"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "lastSearchedAt": {
            "type": "string",
            "nullable": true
          },
          "lastHuntedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "outletId",
          "journalistId",
          "journalistName",
          "firstName",
          "lastName",
          "lastSearchedAt",
          "lastHuntedAt"
        ]
      },
      "NeedHunter": {
        "type": "object",
        "properties": {
          "outletId": {
            "type": "string",
            "format": "uuid"
          },
          "journalistId": {
            "type": "string",
            "format": "uuid"
          },
          "journalistName": {
            "type": "string"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "outletId",
          "journalistId",
          "journalistName",
          "firstName",
          "lastName"
        ]
      },
      "NeedAgentSearch": {
        "type": "object",
        "properties": {
          "outletId": {
            "type": "string",
            "format": "uuid"
          },
          "journalistId": {
            "type": "string",
            "format": "uuid"
          },
          "journalistName": {
            "type": "string"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "outletId",
          "journalistId",
          "journalistName",
          "firstName",
          "lastName"
        ]
      },
      "NeedHunterVerification": {
        "type": "object",
        "properties": {
          "outletId": {
            "type": "string",
            "format": "uuid"
          },
          "journalistId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "outletId",
          "journalistId",
          "email"
        ]
      },
      "JournalistEngagement": {
        "type": "object",
        "properties": {
          "journalistId": {
            "type": "string",
            "format": "uuid"
          },
          "journalistName": {
            "type": "string"
          },
          "pitchBounces": {
            "type": "number"
          },
          "deliveries": {
            "type": "number"
          },
          "lastEngagementAt": {
            "type": "string",
            "nullable": true
          },
          "totalPitches": {
            "type": "number"
          },
          "totalOpens": {
            "type": "number"
          },
          "totalReplies": {
            "type": "number"
          }
        },
        "required": [
          "journalistId",
          "journalistName",
          "pitchBounces",
          "deliveries",
          "lastEngagementAt",
          "totalPitches",
          "totalOpens",
          "totalReplies"
        ]
      },
      "JournalistStatus": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string",
            "format": "uuid"
          },
          "outletId": {
            "type": "string",
            "format": "uuid"
          },
          "journalistId": {
            "type": "string",
            "format": "uuid"
          },
          "journalistName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "relevanceScore": {
            "type": "string"
          }
        },
        "required": [
          "campaignId",
          "outletId",
          "journalistId",
          "journalistName",
          "status",
          "relevanceScore"
        ]
      },
      "JournalistWithEmails": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "journalistName": {
            "type": "string"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "entityType": {
            "type": "string",
            "enum": [
              "individual",
              "organization"
            ]
          },
          "emails": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "isValid": {
                  "type": "boolean"
                },
                "confidence": {
                  "type": "number"
                }
              },
              "required": [
                "email",
                "isValid",
                "confidence"
              ]
            }
          }
        },
        "required": [
          "id",
          "journalistName",
          "firstName",
          "lastName",
          "entityType",
          "emails"
        ]
      }
    },
...
  },
  "paths": {
...
    "/journalists/emails/valid": {
      "get": {
        "summary": "Valid journalist emails",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "outlet_id",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "journalist_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Valid emails",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "emails": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ValidJournalistEmail"
                      }
                    }
                  },
                  "required": [
                    "emails"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/journalists/emails/hunted-events": {
      "get": {
        "summary": "Hunted email events",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "outlet_id",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "journalist_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/HuntedEmailEvent"
                      }
                    }
                  },
                  "required": [
                    "events"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/journalists/emails/hunted-individual-events": {
      "get": {
        "summary": "Hunted individual events",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "outlet_id",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "journalist_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/HuntedIndividualEvent"
                      }
                    }
                  },
                  "required": [
                    "events"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/journalists/emails/searched-events": {
      "get": {
        "summary": "Searched email events",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "outlet_id",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "journalist_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SearchedEmailEvent"
                      }
                    }
                  },
                  "required": [
                    "events"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/journalists/need-email-update": {
      "get": {
        "summary": "Journalists needing email update",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Journalists needing update",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "journalists": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/NeedEmailUpdate"
                      }
                    }
                  },
                  "required": [
                    "journalists"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/journalists/need-hunter": {
      "get": {
        "summary": "Journalists needing Hunter search",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Journalists needing Hunter",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "journalists": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/NeedHunter"
                      }
                    }
                  },
                  "required": [
                    "journalists"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/journalists/need-agent-search": {
      "get": {
        "summary": "Journalists needing agent search",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Journalists needing agent search",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "journalists": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/NeedAgentSearch"
                      }
                    }
                  },
                  "required": [
                    "journalists"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/journalists/emails/need-hunter-verification": {
      "get": {
        "summary": "Emails needing Hunter verification",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Emails needing verification",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "emails": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/NeedHunterVerification"
                      }
                    }
                  },
                  "required": [
                    "emails"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/journalists/engagement/{journalistId}": {
      "get": {
        "summary": "Journalist engagement metrics",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "journalistId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Engagement data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "engagement": {
                      "$ref": "#/components/schemas/JournalistEngagement"
                    }
                  },
                  "required": [
                    "engagement"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/journalists/status": {
      "get": {
        "summary": "Journalist status by campaign",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "campaign_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Statuses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statuses": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/JournalistStatus"
                      }
                    }
                  },
                  "required": [
                    "statuses"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/internal/journalists/by-outlet-with-emails/{outletId}": {
      "get": {
        "summary": "Journalists with valid emails for an outlet",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "outletId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Journalists with emails",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "journalists": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/JournalistWithEmails"
                      }
                    }
                  },
                  "required": [
                    "journalists"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/internal/journalists/by-ids": {
      "get": {
        "summary": "Batch lookup journalists by IDs",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "ids",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Journalists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "journalists": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Journalist"
                      }
                    }
                  },
                  "required": [
                    "journalists"
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}